perf_test: benchmark size as size_t

+++ core/perf_test/PerfTest_ViewCopy.hpp
@@ -150,7 +150,7 @@ static void ViewDeepCopy_Raw(benchmark::State& state) {
 
 template <typename DstMemorySpace, typename SrcMemorySpace>
 static void ViewDeepCopy_Rank1Strided(benchmark::State& state) {
-  const int N8 = std::pow(state.range(0), 8);
+  const size_t N8 = std::pow(state.range(0), 8);
 
   // allocate 2x the size since layout only has 1/2 the elements
   Kokkos::View<double*, DstMemorySpace> a("A1", N8 * 2);
